disables custom metrics

+++ game-server/src/main/java/org/coinen/reactive/pacman/service/support/DefaultPlayerService.java
@@ -138,6 +138,11 @@ public class DefaultPlayerService implements PlayerService {
                            });
 
 
+                           if (updatedPlayer == null) {
+                               return;
+                           }
+
+
                            if (updatedPlayer.getState() == Player.State.DISCONNECTED) {
                                playerRepository.delete(uuid);
                            }
